requirements.txt
python-dotenv=1.1.1
python-telegram-bot=22.4
pydub=0.25.1
ffmpeg-python=0.2.0